+++ lib/interfaces/injectable-class.ts
@@ -2,7 +2,7 @@ import { FastifyInstance } from 'fastify';
 import { CREATOR } from '../symbols';
 
 export interface InjectableClass {
-    [CREATOR]: {
+    [CREATOR]?: {
         register: (instance: FastifyInstance) => void
     }
 }