+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Added
+- User friendly error when trying to bootstrap invalid module
+- Option to skip broken modules to prevent crash when trying to bootstrap invalid module
+
 ## [2.0.0-5]
 ## Added
 - Link to JetBrains IDE plugin (WebStorm, IDEA, ...)